Added Fragment Lighting System
Functionality for ambient, point, directional, and spotlight lighting. Normals are now calculated. Fixed rotater aswell.
