+++ Source/HelpfulScripts/Modulo.h
@@ -1,0 +1,6 @@
+#pragma once
+#include <math.h>
+
+int Modulo(int, int);
+float Modulo(float, float);
+double Modulo(double, double);
