--- Source/Modulo.h
@@ -1,6 +1,0 @@
-#pragma once
-#include <math.h>
-
-int Modulo(int, int);
-float Modulo(float, float);
-double Modulo(double, double);
